+++ packages/jest-circus/src/run.js
@@ -17,7 +17,7 @@ import type {
 
 import {getState, dispatch} from './state';
 import {
-  callAsyncFn,
+  callAsyncCircusFn,
   getAllHooksForDescribe,
   getEachHooksForTest,
   getTestID,
@@ -44,7 +44,7 @@ const _runTestsForDescribeBlock = async (describeBlock: DescribeBlock) => {
   const {beforeAll, afterAll} = getAllHooksForDescribe(describeBlock);
 
   for (const hook of beforeAll) {
-    await _callHook({describeBlock, hook});
+    await _callCircusHook({describeBlock, hook});
   }
 
   // Tests that fail and are retried we run after other tests
@@ -77,7 +77,7 @@ const _runTestsForDescribeBlock = async (describeBlock: DescribeBlock) => {
   }
 
   for (const hook of afterAll) {
-    await _callHook({describeBlock, hook});
+    await _callCircusHook({describeBlock, hook});
   }
   dispatch({describeBlock, name: 'run_describe_finish'});
 };
@@ -105,13 +105,13 @@ const _runTest = async (test: TestEntry): Promise<void> => {
       // hooks after that.
       break;
     }
-    await _callHook({hook, test, testContext});
+    await _callCircusHook({hook, test, testContext});
   }
 
-  await _callTest(test, testContext);
+  await _callCircusTest(test, testContext);
 
   for (const hook of afterEach) {
-    await _callHook({hook, test, testContext});
+    await _callCircusHook({hook, test, testContext});
   }
 
   // `afterAll` hooks should not affect test status (pass or fail), because if
@@ -120,7 +120,7 @@ const _runTest = async (test: TestEntry): Promise<void> => {
   dispatch({name: 'test_done', test});
 };
 
-const _callHook = ({
+const _callCircusHook = ({
   hook,
   test,
   describeBlock,
@@ -133,14 +133,14 @@ const _callHook = ({
 }): Promise<mixed> => {
   dispatch({hook, name: 'hook_start'});
   const timeout = hook.timeout || getState().testTimeout;
-  return callAsyncFn(hook.fn, testContext, {isHook: true, timeout})
+  return callAsyncCircusFn(hook.fn, testContext, {isHook: true, timeout})
     .then(() => dispatch({describeBlock, hook, name: 'hook_success', test}))
     .catch(error =>
       dispatch({describeBlock, error, hook, name: 'hook_failure', test}),
     );
 };
 
-const _callTest = async (
+const _callCircusTest = (
   test: TestEntry,
   testContext: TestContext,
 ): Promise<void> => {
@@ -150,10 +150,10 @@ const _callTest = async (
 
   if (test.errors.length) {
     // We don't run the test if there's already an error in before hooks.
-    return;
+    return Promise.resolve();
   }
 
-  await callAsyncFn(test.fn, testContext, {isHook: false, timeout})
+  return callAsyncCircusFn(test.fn, testContext, {isHook: false, timeout})
     .then(() => dispatch({name: 'test_fn_success', test}))
     .catch(error => dispatch({error, name: 'test_fn_failure', test}));
 };